Change folder structure, merge all sub-packages into "w32" package.
