+++ gdi32.go
@@ -2,10 +2,9 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package gdi32
+package w32
 
 import (
-    . "github.com/AllenDang/w32"
     "syscall"
     "unsafe"
 )
