+++ user32.go
@@ -2,11 +2,10 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package user32
+package w32
 
 import (
     "fmt"
-    . "github.com/AllenDang/w32"
     "syscall"
     "unsafe"
 )
