+++ vars.go
@@ -1,8 +1,4 @@
-package com
-
-import (
-	. "github.com/AllenDang/w32"
-)
+package w32
 
 var (
 	IID_NULL                      = &GUID{0x00000000, 0x0000, 0x0000, [8]byte{0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00}}
